k64ay/todo_infra/requirements.txt
django
djangorestframework
django-filter
markdown
python-dotenv
django-cors-headers
djoser
beautifulsoup4
requests
django-debug-toolbar
gunicorn
celery
redis